backend/requirements.txt
Django>=5,<6
djangorestframework
psycopg2-binary
celery[redis]
pandas
openpyxl
gunicorn
django-cors-headers
